# apps/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kedai Bunda API",
        "description": "API untuk Sistem Kasir dan Manajemen Stok Kedai Bunda",
        "contact": {
            "email": "admin@kedaibunda.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/menu": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Menampilkan daftar menu",
                "operationId": "1a2f6cda46ada094192150c735e84b82",
                "parameters": [
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Filter berdasarkan kategori (makanan/minuman)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "makanan",
                                "minuman"
                            ],
                            "example": "makanan"
                        }
                    },
                    {
                        "name": "tersedia",
                        "in": "query",
                        "description": "Filter menu yang tersedia (1/0)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar menu berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data menu"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Nasi Goreng"
                                                    },
                                                    "kategori": {
                                                        "type": "string",
                                                        "example": "makanan"
                                                    },
                                                    "harga_jual": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 25000
                                                    },
                                                    "gambar": {
                                                        "type": "string",
                                                        "example": "nasi-goreng.jpg"
                                                    },
                                                    "deskripsi": {
                                                        "type": "string",
                                                        "example": "Nasi goreng spesial dengan telur"
                                                    },
                                                    "tersedia": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Menyimpan menu baru",
                "operationId": "637c6c8a2c72ab5ef50d36ec1c61808c",
                "requestBody": {
                    "description": "Data menu baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "kategori",
                                    "harga_jual"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Nasi Goreng"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "enum": [
                                            "makanan",
                                            "minuman"
                                        ],
                                        "example": "makanan"
                                    },
                                    "harga_jual": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25000
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "example": "nasi-goreng.jpg"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Nasi goreng spesial dengan telur"
                                    },
                                    "tersedia": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Menu berhasil ditambahkan"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama": {
                                                    "type": "string",
                                                    "example": "Nasi Goreng"
                                                },
                                                "kategori": {
                                                    "type": "string",
                                                    "example": "makanan"
                                                },
                                                "harga_jual": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 25000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}